Replace int with size_t where possible
Some loop variables were typed to `int` instead of the prefered
`size_t`.

+++ src/twiddle/bloomfilter/bloomfilter.c
@@ -68,7 +68,7 @@ void tw_bloomfilter_set(struct tw_bloomfilter *bf, size_t size, const char *buf)
   struct tw_bitmap *bitmap = bf->bitmap;
   const uint32_t b_size = bitmap->info.size;
 
-  for (int i = 0; i < k; ++i) {
+  for (size_t i = 0; i < k; ++i) {
     tw_bitmap_set(bitmap, (hash.h + i * hash.l) % b_size);
   }
 }
@@ -83,7 +83,7 @@ bool tw_bloomfilter_test(const struct tw_bloomfilter *bf, size_t size,
   const struct tw_bitmap *bitmap = bf->bitmap;
   const uint32_t b_size = bitmap->info.size;
 
-  for (int i = 0; i < k; ++i) {
+  for (size_t i = 0; i < k; ++i) {
     if (!tw_bitmap_test(bitmap, (hash.h + i * hash.l) % b_size)) {
       return false;
     }
